requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.10
crispy-bootstrap5==2024.10
Django==5.1.5
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
elastic-transport==8.17.0
elasticsearch==8.17.1
elasticsearch-dsl==8.17.1
gunicorn==23.0.0
kombu==5.4.2
packaging==24.2
prompt_toolkit==3.0.50
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
